pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "MorganTwoZero", email = "90017582+MorganTwoZero@users.noreply.github.com"},
]
dependencies = [
    "Pillow>=9.2.0",
    "starlite[full]>=1.25.0",
    "httpx>=0.23.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
